+++ src/Products/PloneMeeting/MeetingItem.py
@@ -1419,6 +1419,22 @@ schema = Schema((
         enforceVocabulary=True,
     ),
     TextField(
+        name='emergencyMotivation',
+        widget=RichWidget(
+            condition="python: here.attribute_is_used('emergencyMotivation')",
+            label='EmergencyMotivation',
+            label_msgid='PloneMeeting_label_emergencyMotivation',
+            i18n_domain='PloneMeeting',
+        ),
+        default_content_type="text/html",
+        read_permission="PloneMeeting: Read decision",
+        searchable=False,
+        allowable_content_types=('text/html',),
+        default_output_type="text/x-html-safe",
+        optional=True,
+        write_permission=WriteDecision,
+    ),
+    TextField(
         name='motivation',
         widget=RichWidget(
             condition="python: here.attribute_is_used('motivation')",